Fixed large vert count objects - limit of 4,294,967,295 triangulated vertices

+++ Common/loadOBJ.cpp
@@ -15,8 +15,8 @@ struct PackedVertex{
 
 bool getSimilarVertexIndex(
         PackedVertex & packed,
-        std::map<PackedVertex,unsigned short> & VertexToOutIndex,
-        unsigned short & result
+        std::map<PackedVertex,unsigned int> & VertexToOutIndex,
+        unsigned int & result
         ){
     auto it = VertexToOutIndex.find(packed);
     if (it==VertexToOutIndex.end()) {
@@ -33,16 +33,16 @@ void indexVBO(
         std::vector<glm::vec2> & in_uvs,
         std::vector<glm::vec3> & in_normals,
 
-        std::vector<unsigned short> & out_indices,
+        std::vector<unsigned int> & out_indices,
         std::vector<glm::vec3> & out_vertices,
         std::vector<glm::vec2> & out_uvs,
         std::vector<glm::vec3> & out_normals
         ){
-    std::map<PackedVertex,unsigned short> VertexToOutIndex;
+    std::map<PackedVertex,unsigned int> VertexToOutIndex;
     for (unsigned int i=0; i<in_vertices.size(); i++) {
         PackedVertex packed = {in_vertices[i], in_uvs[i], in_normals[i]};
 
-        unsigned short index;
+        unsigned int index;
         bool found = getSimilarVertexIndex(packed, VertexToOutIndex, index);
         if (found) {
             out_indices.push_back(index);
@@ -51,7 +51,7 @@ void indexVBO(
             out_vertices.push_back(in_vertices[i]);
             out_uvs.push_back(in_uvs[i]);
             out_normals.push_back(in_normals[i]);
-            unsigned short newIndex = (unsigned short)out_vertices.size() - 1;
+            unsigned int newIndex = (unsigned int)out_vertices.size() - 1;
             out_indices.push_back(newIndex);
             VertexToOutIndex[packed] = newIndex;
         }
